refactor: 重构SMTC服务入口
