+++ src/main/index.ts
@@ -28,7 +28,6 @@ import { app, BrowserWindow, globalShortcut } from 'electron';
 import { join } from 'path';
 import { readFileSync, writeFileSync, mkdirSync, existsSync } from 'fs';
 import { exec } from 'child_process';
-import { Worker } from 'worker_threads';
 import { electronApp, optimizer } from '@electron-toolkit/utils';
 import { autoUpdater } from 'electron-updater';
 import { createTray, destroyTray } from './tray';
@@ -61,6 +60,7 @@ import { createHotkeyService } from './services/hotkeyService';
 import { initUpdaterService } from './services/updaterService';
 import { createCaptureWindowService } from './window/captureWindow';
 import { createMainWindowService } from './window/mainWindow';
+import { createSmtcService } from './music/smtcService';
 import {
   hasAnyRunningProcess,
   normalizeProcessName,
@@ -243,9 +243,6 @@ const captureWindowService = createCaptureWindowService({
   getMainWindow: () => mainWindow,
 });
 
-/** SMTC Worker 线程引用 */
-let smtcWorker: Worker | null = null;
-
 /** 灵动岛尺寸常量 */
 const ISLAND_WIDTH = 260;
 const ISLAND_HEIGHT = 42;
@@ -416,85 +413,22 @@ const mainWindowService = createMainWindowService({
 /** 隐藏进程名单轮询计时器 */
 let autoHideProcessWatcher: NodeJS.Timeout | null = null;
 
-/** 记录当前生效的设备ID（仅白名单内程序） */
-let currentDeviceId: string = nowPlayingWhitelist[0] || '';
-
-interface DetectedSourceEntry {
-  isPlaying: boolean;
-  hasTitle: boolean;
-  updatedAt: number;
-}
-
-/** SMTC 原始会话缓存（不受白名单限制），用于获取播放进程按钮 */
-const detectedSourceRuntime = new Map<string, DetectedSourceEntry>();
-
-/** SMTC 白名单内会话运行时缓存（模块级，供 IPC 处理器访问） */
-interface SmtcSessionRuntimeEntry {
-  payload: {
-    title: string;
-    artist: string;
-    album: string;
-    duration_ms: number;
-    position_ms: number;
-    isPlaying: boolean;
-    thumbnail: string | null;
-    canFastForward: boolean;
-    canSkip: boolean;
-    canLike: boolean;
-    canChangeVolume: boolean;
-    canSetOutput: boolean;
-    deviceId: string;
-  };
-  hasTitle: boolean;
-  isPlaying: boolean;
-  playStartedAt: number;
-  updatedAt: number;
-}
-let smtcSessionRuntime: Map<string, SmtcSessionRuntimeEntry> | null = null;
-
 /** 进程隐藏轮询防重入标记 */
 let autoHideCheckInFlight = false;
 
 /** 当前隐藏是否由“隐藏进程命中”触发 */
 let hiddenByAutoHideProcess = false;
 
-/** SMTC 缓存上次回收时间 */
-let lastSmtcCleanupAt = 0;
-
 /** SMTC 自动取消订阅时间（毫秒），0 为永不取消 */
 let smtcUnsubscribeMs = DEFAULT_SMTC_UNSUBSCRIBE_MS;
 
-/** 待确认的播放源切换请求 */
-let pendingSourceSwitchId: string = '';
-let pendingSourceSwitchEntry: SmtcSessionRuntimeEntry | null = null;
-
-/**
- * 检查当前设备ID是否在白名单内
- */
-function isWhitelisted(): boolean {
-  const id = currentDeviceId.toLowerCase();
-  return nowPlayingWhitelist.some(name => id.includes(name.toLowerCase()));
-}
-
-function pickDetectedSourceAppId(): string {
-  let bestPlaying = '';
-  let bestPlayingAt = 0;
-  let bestTitled = '';
-  let bestTitledAt = 0;
-
-  detectedSourceRuntime.forEach((entry, sourceAppId) => {
-    if (entry.isPlaying && entry.updatedAt >= bestPlayingAt) {
-      bestPlaying = sourceAppId;
-      bestPlayingAt = entry.updatedAt;
-    }
-    if (entry.hasTitle && entry.updatedAt >= bestTitledAt) {
-      bestTitled = sourceAppId;
-      bestTitledAt = entry.updatedAt;
-    }
-  });
-
-  return bestPlaying || bestTitled;
-}
+const smtcService = createSmtcService({
+  getMainWindow: () => mainWindow,
+  getWhitelist: () => nowPlayingWhitelist,
+  getSmtcUnsubscribeMs: () => smtcUnsubscribeMs,
+  unsubscribeNeverValue: SMTC_UNSUBSCRIBE_NEVER,
+  cleanupIntervalMs: SMTC_RUNTIME_CLEANUP_INTERVAL_MS,
+});
 
 /**
  * 读取持久化的白名单配置
@@ -678,7 +612,7 @@ const hotkeyService = createHotkeyService({
     });
   },
   onNextSongHotkey: () => {
-    if (!isWhitelisted()) return;
+    if (!smtcService.isWhitelisted()) return;
     sendMediaVirtualKey(0xB0);
   },
   onPlayPauseSongHotkey: () => {
@@ -721,20 +655,14 @@ function registerIpcHandlers(): void {
   registerMediaIpcHandlers({
     getMainWindow: () => mainWindow,
     sendMediaVirtualKey,
-    isWhitelisted,
-    getPendingSourceSwitchId: () => pendingSourceSwitchId,
-    setPendingSourceSwitchId: (id) => {
-      pendingSourceSwitchId = id;
-    },
-    getPendingSourceSwitchEntry: () => pendingSourceSwitchEntry,
-    clearPendingSourceSwitchEntry: () => {
-      pendingSourceSwitchEntry = null;
-    },
-    getCurrentDeviceId: () => currentDeviceId,
-    setCurrentDeviceId: (id) => {
-      currentDeviceId = id;
-    },
-    getSmtcSessionRuntime: () => smtcSessionRuntime,
+    isWhitelisted: smtcService.isWhitelisted,
+    getPendingSourceSwitchId: smtcService.getPendingSourceSwitchId,
+    setPendingSourceSwitchId: smtcService.setPendingSourceSwitchId,
+    getPendingSourceSwitchEntry: smtcService.getPendingSourceSwitchEntry,
+    clearPendingSourceSwitchEntry: smtcService.clearPendingSourceSwitchEntry,
+    getCurrentDeviceId: smtcService.getCurrentDeviceId,
+    setCurrentDeviceId: smtcService.setCurrentDeviceId,
+    getSmtcSessionRuntime: smtcService.getSmtcSessionRuntime,
   });
 
   const writeMainLog = createSessionMainLogger();
@@ -801,7 +729,7 @@ function registerIpcHandlers(): void {
       smtcUnsubscribeMs = value;
     },
     sanitizeSmtcUnsubscribeMs,
-    detectSourceAppId: () => pickDetectedSourceAppId(),
+    detectSourceAppId: smtcService.pickDetectedSourceAppId,
   });
 
   // ===== 歌曲设置 IPC =====
@@ -902,215 +830,6 @@ function registerIpcHandlers(): void {
   });
 }
 
-/**
- * 初始化 SMTC Worker 线程，监听系统媒体信息变更并推送到渲染进程
- * @param win - 主窗口引用
- */
-function initSmtcWorker(win: BrowserWindow | null): void {
-  try {
-    const sessionRuntime = new Map<string, SmtcSessionRuntimeEntry>();
-    smtcSessionRuntime = sessionRuntime;
-
-    const emitCurrentSession = (): void => {
-      if (!win || win.isDestroyed()) return;
-      const currentEntry = currentDeviceId ? sessionRuntime.get(currentDeviceId) : undefined;
-      if (currentEntry?.hasTitle) {
-        win.webContents.send('nowplaying:info', currentEntry.payload);
-      } else {
-        win.webContents.send('nowplaying:info', null);
-      }
-    };
-
-    /** 向渲染进程发送播放源切换请求通知 */
-    const emitSourceSwitchRequest = (sourceAppId: string, title: string, artist: string): void => {
-      if (!win || win.isDestroyed()) return;
-      win.webContents.send('media:source-switch-request', { sourceAppId, title, artist });
-    };
-
-    const workerPath = join(__dirname, 'smtcWorker.js');
-    smtcWorker = new Worker(workerPath);
-
-    smtcWorker.on('message', (msg: {
-      type: string;
-      sourceAppId?: string;
-      session?: {
-        media: { title: string; artist: string; albumTitle: string; thumbnail: string | null } | null;
-        playback: { playbackStatus: number; playbackType: number } | null;
-        timeline: { position: number; duration: number } | null;
-      };
-    }) => {
-      if (!win || win.isDestroyed()) return;
-
-      if (msg.type === 'session-removed') {
-        if (msg.sourceAppId) {
-          detectedSourceRuntime.delete(msg.sourceAppId);
-          sessionRuntime.delete(msg.sourceAppId);
-          if (msg.sourceAppId === pendingSourceSwitchId) {
-            pendingSourceSwitchId = '';
-            pendingSourceSwitchEntry = null;
-          }
-        }
-        if (msg.sourceAppId === currentDeviceId) {
-          currentDeviceId = '';
-          emitCurrentSession();
-        }
-        return;
-      }
-
-      if (msg.type !== 'session-update') return;
-
-      const { sourceAppId = '', session } = msg;
-      const { media, playback, timeline } = session ?? {};
-      const now = Date.now();
-
-      if (now - lastSmtcCleanupAt >= SMTC_RUNTIME_CLEANUP_INTERVAL_MS) {
-        cleanupStaleSmtcRuntime(sessionRuntime);
-        lastSmtcCleanupAt = now;
-      }
-
-      if (sourceAppId) {
-        detectedSourceRuntime.set(sourceAppId, {
-          isPlaying: (playback?.playbackStatus ?? 0) === 4,
-          hasTitle: Boolean(media?.title),
-          updatedAt: now,
-        });
-      }
-
-      const sourceAppIdLower = sourceAppId.toLowerCase();
-      if (!nowPlayingWhitelist.some(name => sourceAppIdLower.includes(name.toLowerCase()))) return;
-
-      const hasTitle = Boolean(media?.title);
-      const isPlaying = (playback?.playbackStatus ?? 0) === 4;
-
-      const payload = {
-        title: media?.title ?? '',
-        artist: media?.artist ?? '',
-        album: media?.albumTitle ?? '',
-        duration_ms: Math.round((timeline?.duration ?? 0) * 1000),
-        position_ms: Math.round((timeline?.position ?? 0) * 1000),
-        isPlaying,
-        thumbnail: media?.thumbnail ?? null,
-        canFastForward: false,
-        canSkip: false,
-        canLike: false,
-        canChangeVolume: false,
-        canSetOutput: false,
-        deviceId: sourceAppId,
-      };
-
-      const prevEntry = sessionRuntime.get(sourceAppId);
-      let playStartedAt = prevEntry?.playStartedAt ?? 0;
-      if (isPlaying) {
-        if (!prevEntry?.isPlaying || playStartedAt <= 0) {
-          playStartedAt = Date.now();
-        }
-      } else {
-        playStartedAt = 0;
-      }
-
-      sessionRuntime.set(sourceAppId, {
-        payload,
-        hasTitle,
-        isPlaying,
-        playStartedAt,
-        updatedAt: now,
-      });
-
-      // ===== 播放源锁定逻辑 =====
-
-      // 尚未锁定任何播放源：第一个在播放且有标题的白名单源自动锁定
-      if (!currentDeviceId) {
-        if (isPlaying && hasTitle) {
-          currentDeviceId = sourceAppId;
-          emitCurrentSession();
-        }
-        return;
-      }
-
-      // 来自当前锁定源的更新：直接推送
-      if (sourceAppId === currentDeviceId) {
-        emitCurrentSession();
-        // 如果锁定源停止播放且无标题，解锁
-        if (!isPlaying && !hasTitle) {
-          currentDeviceId = '';
-        }
-        return;
-      }
-
-      // 来自其他源的更新：如果它正在播放且有标题
-      if (isPlaying && hasTitle) {
-        const lockedEntry = sessionRuntime.get(currentDeviceId);
-        if (lockedEntry?.isPlaying) {
-          // 锁定源正在播放 → 弹通知询问是否切换
-          if (pendingSourceSwitchId === sourceAppId) {
-            pendingSourceSwitchEntry = sessionRuntime.get(sourceAppId) ?? null;
-            return;
-          }
-          pendingSourceSwitchId = sourceAppId;
-          pendingSourceSwitchEntry = sessionRuntime.get(sourceAppId) ?? null;
-          emitSourceSwitchRequest(sourceAppId, payload.title, payload.artist);
-        } else {
-          // 锁定源未在播放 → 直接切换到新源，无需通知
-          currentDeviceId = sourceAppId;
-          emitCurrentSession();
-        }
-      }
-    });
-
-    smtcWorker.on('error', (err) => {
-      console.error('[SMTC] Worker error:', err);
-    });
-
-    smtcWorker.on('exit', (code) => {
-      if (code !== 0) console.error('[SMTC] Worker exited with code:', code);
-    });
-  } catch (err) {
-    console.error('[SMTC] Worker init error:', err);
-  }
-}
-
-/**
- * 终止 SMTC Worker 线程并释放资源
- */
-function cleanupSmtcWorker(): void {
-  if (smtcWorker) {
-    smtcWorker.terminate();
-    smtcWorker = null;
-  }
-  detectedSourceRuntime.clear();
-  smtcSessionRuntime?.clear();
-  smtcSessionRuntime = null;
-  pendingSourceSwitchId = '';
-  pendingSourceSwitchEntry = null;
-  currentDeviceId = '';
-  lastSmtcCleanupAt = 0;
-}
-
-function cleanupStaleSmtcRuntime(sessionRuntime: Map<string, SmtcSessionRuntimeEntry>): void {
-  const now = Date.now();
-  const ttlMs = smtcUnsubscribeMs;
-  if (ttlMs === SMTC_UNSUBSCRIBE_NEVER) return;
-
-  detectedSourceRuntime.forEach((entry, sourceAppId) => {
-    if (now - entry.updatedAt > ttlMs) {
-      detectedSourceRuntime.delete(sourceAppId);
-    }
-  });
-
-  sessionRuntime.forEach((entry, sourceAppId) => {
-    if (now - entry.updatedAt > ttlMs) {
-      sessionRuntime.delete(sourceAppId);
-      if (sourceAppId === pendingSourceSwitchId) {
-        pendingSourceSwitchId = '';
-        pendingSourceSwitchEntry = null;
-      }
-      if (sourceAppId === currentDeviceId) {
-        currentDeviceId = '';
-      }
-    }
-  });
-}
-
 // ===== 剪贴板 URL 监听 =====
 
 /** 剪贴板 URL 监听是否启用 */
@@ -1137,7 +856,7 @@ registerAppLifecycleHandlers({
   },
   onWindowAllClosed: () => {
     stopAutoHideProcessWatcher();
-    cleanupSmtcWorker();
+    smtcService.cleanupWorker();
     destroyTray();
     if (process.platform !== 'darwin') {
       app.quit();
@@ -1163,7 +882,7 @@ app.whenReady().then(() => {
   mainWindowService.createWindow();
   createTray(mainWindow);
 
-  initSmtcWorker(mainWindow);
+  smtcService.initWorker();
   startClipboardUrlWatcher({
     getWindow: () => mainWindow,
     getEnabled: () => clipboardUrlMonitorEnabled,
